Queue new batches for indexing
